requirements.txt
aiogram
aiohttp
requests
python-dotenv==1.0.0
openai==0.27.8
asyncpg

